fix slave index

+++ src/s_sync.c
@@ -61,45 +61,45 @@ uint8_t getPacketType(uint8_t* packet){
 	return packet[1];
 }
 
-void setSyncTimeStamp(uint8_t* packet, uint64_t timestamp){
-	packet[2+0] = uint8_t((timestamp >> 8*7) & 0xFF);
-	packet[2+1] = uint8_t((timestamp >> 8*6) & 0xFF);
-	packet[2+2] = uint8_t((timestamp >> 8*5) & 0xFF);
-	packet[2+3] = uint8_t((timestamp >> 8*4) & 0xFF);
-	packet[2+4] = uint8_t((timestamp >> 8*3) & 0xFF);
-	packet[2+5] = uint8_t((timestamp >> 8*2) & 0xFF);
-	packet[2+6] = uint8_t((timestamp >> 8*1) & 0xFF);
-	packet[2+7] = uint8_t((timestamp >> 8*0) & 0xFF);
+void setSyncTimeStamp(uint8_t* packet, int64_t timestamp){
+	packet[2+0] = (timestamp >> 8*7) & 0xFF;
+	packet[2+1] = (timestamp >> 8*6) & 0xFF;
+	packet[2+2] = (timestamp >> 8*5) & 0xFF;
+	packet[2+3] = (timestamp >> 8*4) & 0xFF;
+	packet[2+4] = (timestamp >> 8*3) & 0xFF;
+	packet[2+5] = (timestamp >> 8*2) & 0xFF;
+	packet[2+6] = (timestamp >> 8*1) & 0xFF;
+	packet[2+7] = (timestamp >> 8*0) & 0xFF;
 
 }
 
-uint64_t getSyncTimeStamp(uint8_t* packet){
-	uint64_t timestamp;
-	timestamp  = packet[2+0] << 8*7;
-	timestamp |= packet[2+1] << 8*6;
-	timestamp |= packet[2+2] << 8*5;
-	timestamp |= packet[2+3] << 8*4;
-	timestamp |= packet[2+4] << 8*3;
-	timestamp |= packet[2+5] << 8*2;
-	timestamp |= packet[2+6] << 8*1;
-	timestamp |= packet[2+7] << 8*0;
+int64_t getSyncTimeStamp(uint8_t* packet){
+	int64_t timestamp;
+	timestamp  = (uint64_t)packet[2+0] << 8*7;
+	timestamp |= (uint64_t)packet[2+1] << 8*6;
+	timestamp |= (uint64_t)packet[2+2] << 8*5;
+	timestamp |= (uint64_t)packet[2+3] << 8*4;
+	timestamp |= (uint64_t)packet[2+4] << 8*3;
+	timestamp |= (uint64_t)packet[2+5] << 8*2;
+	timestamp |= (uint64_t)packet[2+6] << 8*1;
+	timestamp |= (uint64_t)packet[2+7] << 8*0;
 	return timestamp;
 }
 
-void setDelayResp(uint8_t* packet, uint32_t timestamp, uint8_t slave_index){
+void setDelayResp(uint8_t* packet, int32_t timestamp, uint8_t slave_index){
 	uint8_t index = slave_index * 4 + 2;
-	packet[index+0] = uint8_t((timestamp >> 4*3) & 0xFF);
-	packet[index+1] = uint8_t((timestamp >> 4*2) & 0xFF);
-	packet[index+2] = uint8_t((timestamp >> 4*1) & 0xFF);
-	packet[index+3] = uint8_t((timestamp >> 4*0) & 0xFF);
+	packet[index+0] = (timestamp >> 4*3) & 0xFF;
+	packet[index+1] = (timestamp >> 4*2) & 0xFF;
+	packet[index+2] = (timestamp >> 4*1) & 0xFF;
+	packet[index+3] = (timestamp >> 4*0) & 0xFF;
 }
 
-uint32_t getDelayResp(uint8_t* packet, uint8_t slave_index){
+int32_t getDelayResp(uint8_t* packet, uint8_t slave_index){
 	uint8_t index = slave_index * 4 + 2;
-	uint32_t timestamp;
-	timestamp  = packet[index+0] << 4*3;
-	timestamp |= packet[index+1] << 4*2;
-	timestamp |= packet[index+2] << 4*1;
-	timestamp |= packet[index+3] << 4*0;
+	int32_t timestamp;
+	timestamp  = (uint32_t)packet[index+0] << 4*3;
+	timestamp |= (uint32_t)packet[index+1] << 4*2;
+	timestamp |= (uint32_t)packet[index+2] << 4*1;
+	timestamp |= (uint32_t)packet[index+3] << 4*0;
 	return timestamp;
 }
